order / order item DTO, Data Type changes

+++ backend/order-service/src/main/java/com/buyone/orderservice/dto/response/OrderResponse.java
@@ -3,6 +3,8 @@ package com.buyone.orderservice.dto.response;
 import com.buyone.orderservice.model.OrderStatus;
 import lombok.Builder;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -11,7 +13,7 @@ import java.util.List;
 public class OrderResponse {
     private String orderNumber;
     private OrderStatus status;
-    private double total;
+    private BigDecimal total;
     private LocalDateTime createdAt;
     private List<OrderItemResponse> items;  // Nested for details
 }
